Don't use BLKBSZSET on Linux.
This was a workaround for the previous lack of buffering when
newfs_msdos would write in 512 byte sectors even if the underlying block
device was doing 1024 bytes. Now we have a 1MiB buffer, overriding the
sector size shouldn't make a difference.

This brings us closer in line with upstream, which never had this change.

Test: treehugger
Change-Id: I51f9a46b1d9418820b938541adb7e8bc22ac46b5